Review Changes: refacto speaker infos + renamed day to reccuringDay
