+++ remotion/compositions/templates/talk/Talks.composition.tsx
@@ -42,7 +42,7 @@ export const TalksComposition: React.FC = () => {
 					title: 'Certification “Google Cloud Architect”',
 					startingDate: new Date(2023, 3, 18, 13),
 					endingDate: new Date(2023, 4, 23, 13, 45),
-					day: 'mardi',
+					reccuringDay: 'mardi',
 					logoUrl:
 						'https://user-images.githubusercontent.com/72607059/233019842-047a34a4-77c1-4200-adc8-c70a6daf8f10.svg',
 					speaker: {
